post id 필드에 snapshot id가 들어가게끔 해서 id가 없는 상황 방지
